pull yAxisSplit out of LineAreaBarRenderer and call from settings
