+++ frontend/src/metabase/visualizations/lib/settings/graph.js
@@ -22,6 +22,7 @@ import { columnSettings } from "metabase/visualizations/lib/settings/column";
 import { getOptionFromColumn } from "metabase/visualizations/lib/settings/utils";
 import { dimensionIsNumeric } from "metabase/visualizations/lib/numeric";
 import { dimensionIsTimeseries } from "metabase/visualizations/lib/timeseries";
+import getYAxisSplit from "metabase/visualizations/lib/yAxisSplit";
 
 import _ from "underscore";
 
@@ -510,6 +511,9 @@ export const GRAPH_AXIS_SETTINGS = {
     widget: "toggle",
     default: true,
     getHidden: series => series.length < 2,
+  },
+  "graph.y_axis._split_indexes": {
+    getDefault: getYAxisSplit,
   },
   "graph.x_axis.labels_enabled": {
     section: t`Labels`,
